requirements.txt
paramiko==2.8.1
python-dotenv
pandas
google-cloud-bigquery
pyarrow
pandas-gbq
numpy